+++ CHANGES.rst
@@ -27,10 +27,11 @@ New features
 
 Major changes
 -------------
-* The :class:`TextEncoder` is now available to encode columns with diverse entries.
+* The :class:`TextEncoder` is now available to encode string columns with
+  diverse entries.
   It allows the representation of table entries as embeddings computed by a deep
-  learning model. The weights of this model can be fetched locally or from the
-  HuggingFace Hub.
+  learning language model. The weights of this model can be fetched locally
+  or from the HuggingFace Hub.
   :pr:`1077` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
 
 Minor changes